Automatically backed up by Learn

+++ db/migrate/01_create_users.rb
@@ -4,7 +4,7 @@ class CreateUsers < ActiveRecord::Migration[5.0]
     create_table :users do |t|
       t.string :username
       t.string :password
-      t.integer :balance
+      t.decimal :balance
 
     end
   end
